Update gnu-efi to v4.0.0
shim comes with a vendored copy of gnu-efi, which is based on
upstream v3.0.12 with a significant number of changes on top,
some of which are backported from upstream and some of which
are specific to shim.

Unfortunately this version of gnu-efi doesn't come with
riscv64 support, so we can't use to build shim on riscv64;
fortunately though, it is possible to just swap in the
upstream v4.0.0 release to solve the issue.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>